refactor: 대여 예약 검증 서비스과 대여 예약 조회 서비스 분리
