+++ src/main/java/com/girigiri/kwrental/reservation/service/ReservationService.java
@@ -3,13 +3,9 @@ package com.girigiri.kwrental.reservation.service;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.YearMonth;
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
@@ -18,8 +14,6 @@ import org.springframework.transaction.annotation.Transactional;
 import com.girigiri.kwrental.reservation.domain.EquipmentReservationWithMemberNumber;
 import com.girigiri.kwrental.reservation.domain.LabRoomReservation;
 import com.girigiri.kwrental.reservation.domain.Reservation;
-import com.girigiri.kwrental.reservation.domain.ReservationCalendar;
-import com.girigiri.kwrental.reservation.domain.ReservationSpec;
 import com.girigiri.kwrental.reservation.domain.ReservationSpecStatus;
 import com.girigiri.kwrental.reservation.dto.request.AddEquipmentReservationRequest;
 import com.girigiri.kwrental.reservation.dto.request.AddLabRoomReservationRequest;
@@ -34,27 +28,21 @@ import com.girigiri.kwrental.reservation.dto.response.UnterminatedEquipmentReser
 import com.girigiri.kwrental.reservation.dto.response.UnterminatedLabRoomReservationsResponse;
 import com.girigiri.kwrental.reservation.exception.NotSameRentableRentException;
 import com.girigiri.kwrental.reservation.exception.ReservationNotFoundException;
-import com.girigiri.kwrental.reservation.exception.ReservationSpecException;
 import com.girigiri.kwrental.reservation.repository.ReservationRepository;
 import com.girigiri.kwrental.reservation.repository.ReservationSpecRepository;
 
+import lombok.RequiredArgsConstructor;
+
 @Service
+@RequiredArgsConstructor
 public class ReservationService {
 
 	private final ReservationRepository reservationRepository;
 	private final ReservationSpecRepository reservationSpecRepository;
 	private final ReservationCancelService reservationCancelService;
 	private final ReservationReserveService reservationReserveService;
-
-	public ReservationService(final ReservationRepository reservationRepository,
-		final ReservationSpecRepository reservationSpecRepository,
-		final ReservationCancelService reservationCancelService,
-		final ReservationReserveService reservationReserveService) {
-		this.reservationRepository = reservationRepository;
-		this.reservationSpecRepository = reservationSpecRepository;
-		this.reservationCancelService = reservationCancelService;
-		this.reservationReserveService = reservationReserveService;
-	}
+	private final ReservationRetrieveService reservationRetrieveService;
+	private final ReservationValidateService reservationValidateService;
 
 	@Transactional
 	public void reserveEquipment(final Long memberId, final AddEquipmentReservationRequest addReservationRequest) {
@@ -69,45 +57,19 @@ public class ReservationService {
 	@Transactional(readOnly = true)
 	public ReservationsByEquipmentPerYearMonthResponse getReservationsByEquipmentsPerYearMonth(final Long equipmentId,
 		final YearMonth yearMonth) {
-		LocalDate startOfMonth = yearMonth.atDay(1);
-		LocalDate endOfMonth = yearMonth.atEndOfMonth();
-		final List<ReservationSpec> reservationSpecs = reservationSpecRepository.findNotCanceldByStartDateBetween(
-			equipmentId, startOfMonth, endOfMonth);
-		final ReservationCalendar calendar = ReservationCalendar.from(startOfMonth, endOfMonth);
-		calendar.addAll(reservationSpecs);
-		return ReservationsByEquipmentPerYearMonthResponse.from(calendar);
+		return reservationRetrieveService.getReservationsByEquipmentsPerYearMonth(equipmentId, yearMonth);
 	}
 
 	@Transactional(readOnly = true, propagation = Propagation.MANDATORY)
 	public Set<EquipmentReservationWithMemberNumber> getReservationsByStartDate(final LocalDate startDate) {
-		return reservationSpecRepository.findEquipmentReservationWhenAccept(startDate);
+		return reservationRetrieveService.getReservationsByStartDate(startDate);
 	}
 
 	@Transactional(readOnly = true, propagation = Propagation.MANDATORY)
 	public Map<Long, Set<String>> validatePropertyNumbersCountAndGroupByEquipmentId(final Long reservationId,
 		final Map<Long, Set<String>> propertyNumbersByReservationSpecId) {
-		final Reservation reservation = reservationRepository.findByIdWithSpecs(reservationId)
-			.orElseThrow(ReservationNotFoundException::new);
-		validateReservationSpecIdContainsAll(reservation, propertyNumbersByReservationSpecId.keySet());
-		Map<Long, Set<String>> collectedByEquipmentId = new HashMap<>();
-		for (ReservationSpec reservationSpec : reservation.getReservationSpecs()) {
-			reservationSpec.validateAmount(propertyNumbersByReservationSpecId.get(reservationSpec.getId()).size());
-			collectedByEquipmentId.put(reservationSpec.getRentable().getId(),
-				propertyNumbersByReservationSpecId.get(reservationSpec.getId()));
-		}
-		return collectedByEquipmentId;
-	}
-
-	private void validateReservationSpecIdContainsAll(final Reservation reservation,
-		final Set<Long> reservationSpecIdsFromInput) {
-		final Set<Long> reservationSpecIdsFromReservation = reservation.getReservationSpecs().stream()
-			.filter(ReservationSpec::isReserved)
-			.map(ReservationSpec::getId)
-			.collect(Collectors.toSet());
-		if (reservationSpecIdsFromInput.containsAll(reservationSpecIdsFromReservation) &&
-			reservationSpecIdsFromReservation.containsAll(reservationSpecIdsFromInput))
-			return;
-		throw new ReservationSpecException("신청한 대여 상세와 입력된 대여 상세가 일치하지 않습니다.");
+		return reservationValidateService.validatePropertyNumbersCountAndGroupByEquipmentId(reservationId,
+			propertyNumbersByReservationSpecId);
 	}
 
 	@Transactional(propagation = Propagation.MANDATORY)
@@ -125,37 +87,28 @@ public class ReservationService {
 
 	@Transactional(readOnly = true, propagation = Propagation.MANDATORY)
 	public Set<EquipmentReservationWithMemberNumber> getOverdueReservationsWithMemberNumber(final LocalDate localDate) {
-		return reservationSpecRepository.findOverdueEquipmentReservationWhenReturn(localDate);
+		return reservationRetrieveService.getOverdueReservationsWithMemberNumber(localDate);
 	}
 
 	@Transactional(readOnly = true, propagation = Propagation.MANDATORY)
 	public Set<EquipmentReservationWithMemberNumber> getReservationsWithMemberNumberByEndDate(
 		final LocalDate localDate) {
-		return reservationSpecRepository.findEquipmentReservationWhenReturn(localDate);
+		return reservationRetrieveService.getReservationsWithMemberNumberByEndDate(localDate);
 	}
 
 	@Transactional(readOnly = true, propagation = Propagation.MANDATORY)
 	public Reservation getReservationWithReservationSpecsById(final Long id) {
-		return reservationRepository.findByIdWithSpecs(id)
-			.orElseThrow(ReservationNotFoundException::new);
+		return reservationRetrieveService.getReservationWithReservationSpecsById(id);
 	}
 
 	@Transactional(readOnly = true)
 	public UnterminatedEquipmentReservationsResponse getUnterminatedEquipmentReservations(final Long memberId) {
-		final Set<Reservation> reservations = reservationRepository.findNotTerminatedEquipmentReservationsByMemberId(
-			memberId);
-		final List<Reservation> reservationList = new ArrayList<>(reservations);
-		reservationList.sort(Comparator.comparing(Reservation::getRentalPeriod));
-		return UnterminatedEquipmentReservationsResponse.from(reservationList);
+		return reservationRetrieveService.getUnterminatedEquipmentReservations(memberId);
 	}
 
 	@Transactional(readOnly = true)
 	public UnterminatedLabRoomReservationsResponse getUnterminatedLabRoomReservations(final Long memberId) {
-		final Set<Reservation> reservations = reservationRepository.findNotTerminatedLabRoomReservationsByMemberId(
-			memberId);
-		final List<Reservation> reservationList = new ArrayList<>(reservations);
-		reservationList.sort(Comparator.comparing(Reservation::getRentalPeriod));
-		return UnterminatedLabRoomReservationsResponse.from(reservationList);
+		return reservationRetrieveService.getUnterminatedLabRoomReservations(memberId);
 	}
 
 	@Transactional
@@ -165,12 +118,12 @@ public class ReservationService {
 
 	@Transactional(readOnly = true)
 	public Set<LabRoomReservationWithMemberNumberResponse> getLabRoomReservationForAccept(final LocalDate date) {
-		return reservationSpecRepository.findLabRoomReservationsWhenAccept(date);
+		return reservationRetrieveService.getLabRoomReservationForAccept(date);
 	}
 
 	@Transactional(readOnly = true)
 	public Set<LabRoomReservationWithMemberNumberResponse> getLabRoomReservationForReturn(final LocalDate date) {
-		return reservationSpecRepository.findLabRoomReservationWhenReturn(date);
+		return reservationRetrieveService.getLabRoomReservationForReturn(date);
 	}
 
 	@Transactional(propagation = Propagation.MANDATORY)
@@ -212,21 +165,17 @@ public class ReservationService {
 
 	@Transactional(readOnly = true)
 	public RelatedReservationsInfoResponse getRelatedReservationsInfo(Long id) {
-		Reservation reservation = getReservationById(id);
-		LabRoomReservation labRoomReservation = new LabRoomReservation(reservation);
-		final List<Reservation> reservations = reservationRepository.findNotTerminatedRelatedReservation(
-			labRoomReservation);
-		return RelatedReservationsInfoResponse.from(reservations);
+		return reservationRetrieveService.getRelatedReservationsInfo(id);
 	}
 
 	@Transactional(readOnly = true)
 	public HistoryStatResponse getHistoryStat(String name, LocalDate startDate, LocalDate endDate) {
-		return reservationSpecRepository.findHistoryStat(name, startDate, endDate);
+		return reservationRetrieveService.getHistoryStat(name, startDate, endDate);
 	}
 
 	@Transactional(readOnly = true)
 	public ReservationPurposeResponse getPurpose(Long id) {
-		return new ReservationPurposeResponse(getReservationById(id).getPurpose());
+		return reservationRetrieveService.getPurpose(id);
 	}
 
 	@Transactional(propagation = Propagation.MANDATORY)
